+++ src/slices/_index.ts
@@ -1,0 +1,6 @@
+export { default as aiRecommendReducer } from './aiRecommendSlice';
+export { default as artworkReducer } from './artworkSlice';
+export { default as collectionReducer } from './collectionSlice';
+export { default as coverReducer } from './coverSlice';
+export { default as exhibitReducer } from './exhibitSlice';
+export { default as themeReducer } from './themeSlice';
